refactor(frontend): extract ChipLegend component and derive CHIP_CATEGORIES from STYLES (#1225)
Address architecture review notes: ChipLegend moved to components/corrections/ per
co-location convention; CHIP_CATEGORIES derived from STYLES record in correction-colors.ts
so new categories auto-include in the legend.
